Split main, drop namedtuples, add type hints
