+++ prices/utils.pyi
@@ -1,0 +1,6 @@
+from typing import Sequence, TypeVar
+
+T = TypeVar('T')
+
+
+def sum(values: Sequence[T]) -> T: ...
